Added player, map, camera follow
Added the player and infinite map with placeholder sprites and set up the player movement.

Also added cinemachine addon that allows for easy player following with the in-game camera instead of scripting it.

Compiled the game into a testable .exe file you can run outside the game located in the "Test" subfolder of the Git, just launch the .exe