Added health potion item effect
Added the first store item- a health potion
Only way to enable it currently without a store is setting the bool of it being purchased to on manually. Regenerates 5HP every 2 seconds up to maximum.

Also added small indication of it being purchased in the form of a minecraft health potion appearing next to the health bar, animation pending.